# LIDAR point cloud generation optimisation