--- a/src/webots/nodes/WbLidar.cpp
+++ b/src/webots/nodes/WbLidar.cpp
@@ -375,7 +375,7 @@ void WbLidar::updatePointCloud(int minWidth, int maxWidth) {
 
   const double dphi = (numberOfLayer > 1) ? (-verticalFieldOfView() / (numberOfLayer - 1)) : 0.0;
   const double cosdPhi = cos(dphi), sindPhi = sin(dphi);
-  const double phi0 = ((numberOfLayer > 1) ? (verticalFieldOfView() / 2) : 0.0) + mCurrentTiltAngle;
+  const double phi0 = ((numberOfLayers > 1) ? (verticalFieldOfView() / 2) : 0.0) + mCurrentTiltAngle;
   const double cosPhi0 = cos(phi0), sinPhi0 = sin(phi0);
 
   const double dtheta = mIsActuallyRotating ? (-2 * M_PI / (double)resolution) : (-actualFieldOfView() / (w - 1.0));