--- a/src/webots/nodes/WbLidar.cpp
+++ b/src/webots/nodes/WbLidar.cpp
@@ -382,7 +382,8 @@ void WbLidar::updatePointCloud(int minWidth, int maxWidth) {
   const double cosdTheta = cos(dtheta);
   const double sindTheta = sin(dtheta);
   const double theta0 = mIsActuallyRotating ? (minWidth * dtheta) : (actualFieldOfView() / 2 + minWidth * dtheta);
-  const double cosTheta0 = cos(theta0), sinTheta0 = sin(theta0);
+  const double cosTheta0 = cos(theta0);
+  const double sinTheta0 = sin(theta0);
 
   // cos(x+dx) = cos(x)cos(dx)-sin(x)sin(dx)
   // sin(x+dx) = sin(x)cos(dx)+cos(x)sin(dx)