--- a/src/webots/nodes/WbLidar.cpp
+++ b/src/webots/nodes/WbLidar.cpp
@@ -374,7 +374,8 @@ void WbLidar::updatePointCloud(int minWidth, int maxWidth) {
   const double t0 = WbSimulationState::instance()->time() / 1000.0 + minWidth * dt;
 
   const double dphi = (numberOfLayers > 1) ? (-verticalFieldOfView() / (numberOfLayers - 1)) : 0.0;
-  const double cosdPhi = cos(dphi), sindPhi = sin(dphi);
+  const double cosdPhi = cos(dphi);
+  const double sindPhi = sin(dphi);
   const double phi0 = ((numberOfLayers > 1) ? (verticalFieldOfView() / 2) : 0.0) + mCurrentTiltAngle;
   const double cosPhi0 = cos(phi0);
   const double sinPhi0 = sin(phi0);