**Description**
After having pinpointed by @lukicdarkoo in https://github.com/cyberbotics/webots/issues/3476
The `WbLidar::updatePointCloud` method is probably the biggest bottleneck (by far) in the LiDAR implementation:
https://github.com/cyberbotics/webots/blob/c5637734a664733a585219c178eb753c0ff8f37b/src/webots/nodes/WbLidar.cpp#L354-L381

**Details**
The optimization is made using three way : 
 1) Factorize code that could be, to avoid re-computations. In particular find a way to move the condition like `if (mIsActuallyRotating)` outside the loop.
 2) Use increment instead of multiplication : replace values of the form (expression) + (counter) * (increment). We initialize the variable to an initial one, and then add at each loop a fixed amount.
 3) Replace cos() and sin() usage by an iterative method : 
Considering that we are taking cos and sin of angle that are only incremented by a fixed amount, we could use addition law on cos and sin as : 
 cos(x+dx) = cos(x)cos(dx)-sin(x)sin(dx)
 sin(x+dx) = sin(x)cos(dx)+cos(x)sin(dx)
with x is the previous value angle, and dx the angle increment.
this way , we could compute the next cos/sin that will be used on the next loop, using the current values, and the cos/sin of the increment, which is actually fixed as well. Thus cos and sin computation is only 2 multiplication and 1 addition each.

**Results**
First, I've checked as I could, so I hope I've didn't introduce bugs or miscalculations. Also, the way I compute variable (iterative way) could introduce some numerical error, but using `double` should be enough.
Execution speed is obtained by using the Time_uclock function, which will obviously be removed before merging. We record the number of us per function call, and also the number of ns per points computed.
Result should be compared using the branch origin/enhancement-optimize-lidar-point-cloud-initial, which include the original code, and add those timing things.
Maybe Webots use to do those profiling, but I found to be the most straight forward way, let me know if their are "official stuff for that.

Here is what I end up with : here are the ns per points over about 100.

  | initial | optimized
-- | -- | --
min | 14.14 | 2.31
max | 21.74 | 2.88
average | 18.85 | 2.47

The way 1) give minor improvement, but enable 2) which made the timing drop from 20ns to 15ns, and most important enable 3) which bring time from 15ns to about 2ns.

So about 7.5 times faster :) 

**Setup**
<details>
<summary>World</summary>
<pre>
#VRML_SIM R2021c utf8
WorldInfo {
  basicTimeStep 20
  FPS 25
  coordinateSystem "NUE"
}
Viewpoint {
  orientation -0.6931032800836722 0.6931032800836722 0.1980295085953349 0.75
  position 1.2 1.6 2.3
}
TexturedBackground {
}
TexturedBackgroundLight {
}
RectangleArena {
}
Robot {
  children [
    Lidar {
      horizontalResolution 3587
      verticalFieldOfView 0.5236
      numberOfLayers 600
      type "rotating"
    }
  ]
  controller "my_controller"
}
</pre>
</details>


<details>
<summary>Controller</summary>
<pre>
#include <webots/robot.h>
#include <webots/lidar.h>

#define TIME_STEP 25

int main(int argc, char **argv) {
  wb_robot_init();

  WbDeviceTag lidar = wb_robot_get_device("lidar");
  wb_lidar_enable(lidar, TIME_STEP);
  wb_lidar_enable_point_cloud(lidar);

  while (wb_robot_step(TIME_STEP) != -1) {
    wb_lidar_get_point_cloud(lidar);
  };

  wb_robot_cleanup();
  return 0;
}
</pre>
</details>