@ShuffleWire 2021-07-26
> > So about 7.5 times faster :)
> 
> Quite impressive!
> 
> Did you compare it in the debug or release mode? It might be a good idea to compare it in the release mode as the compiler does many optimizations as well.

I did it in release mode only :)