@mael25 2021-07-26
> > > So about 7.5 times faster :)
> > 
> > 
> > Quite impressive!
> > Did you compare it in the debug or release mode? It might be a good idea to compare it in the release mode as the compiler does many optimizations as well.
> 
> Thanks :)
> I did it in release mode only (regular make -j18, I'm right ?)

Yes this is correct.

Thank you for your contribution, it is a good idea to use the addition law of cos and sin to improve the performances. I have only a few small changes requested regarding our coding conventions (see below). 

Also, it seems that you do not have the latest version of `webots_ros` in your PR, you should be able to update it by doing `git pull --recurse-submodules`.