@mael25 2021-07-26
I let you also, as you mentioned, remove the speed measurements so we can have a final version and see if the tests pass!