@mael25 2021-07-26
You can try to do:
```
git checkout develop
git pull --recurse-submodules
git checkout enhancement-optimize-lidar-point-cloud
git merge develop
git push
```