@mael25 2021-07-26
It was needed in fact to do `git pull --recurse-submodules` in `develop`, and then checkout this PR and add the changes,