@lukicdarkoo 2021-07-28
Very clever optimization! It misses a changelog though, users should know about this :)