@ShuffleWire 2021-07-28
@lukicdarkoo the changelog section for 2021c doesn't seems ready, should I make it ?
Anyway, I think I will find time to work more on the lidar, I will add an entry to the changelog then :)